+++ src/main/java/com/vm/request/NewQuestionObject.java
@@ -14,7 +14,7 @@ import java.util.List;
 @NoArgsConstructor
 public class NewQuestionObject {
     private Long questionId;
-    private Long surveyId;
+    private Integer surveyId;
     private String questionText;
     private Integer questionTypeId;
     private String responseFormat;
